fix: Moved some utility into categorized folders
